backend/requirements.txt
flake8==6.0.0
peewee==3.16.3
python-telegram-bot==13.11
fastapi==0.103.2
python-jose==3.3.0
uvicorn==0.23.2
psycopg2-binary==2.9.9
